Common Mistakes in Python
This is a IPython Notebook with code snippets to explain the common mistakes which python encounters during programming.
I've tried to explain the mistakes as simple as possible what i have understand. I feel like knowledge of different type of programming errors takes a programmer to next level. 
Definitely, this list is not complete. It will be updated and would be great more people contribute towards it.